Factor RightSidebar and SidebarTabs out of EditorRightSidebar for reuse
